Separate file for LayerType
